Remove statement::execute() in favour of database::execute()
It makes more semantic sense for the database to execute the
statement rather than the statement executing itself. This also
has the knock-on effect of turning statement into a value object
and removing sqlite::make_database. Non-copyablilty of database
with the optional use of std::make_unique/shared is more flexible,
allowing pure stack based use of the library.